sample/main/requirements.txt
ansicolors>=1.1.8  # Imports the 'colors' package
